feat: Improve pixel display with red border for specific coordinates
The code changes add a red border to the pixels at coordinates (46*size, y) and (94*size, y), as well as (x, 46*size) and (x, 94*size). This enhancement improves the visibility of these pixels and makes them stand out.

The recent user commits and repository commits indicate that the initial HTML structure was added to index.html.